refactor #5 - extract out delete view
